src/TradingAssistant/requirements.txt
akshare==0.7.7
jqdatasdk==1.8.9
schedule==1.0.0
PyMySQL==1.0.2
SQLAlchemy==1.3.20
pandas==1.1.3
PyYAML==5.4.1

mysql-connector-python~=8.0.22
email~=6.0.0a1
yaml~=0.2.5
cacheout~=0.13.1